Refactoring + improved sheets model
