+++ src/app/CompetitionBoard/CompetitionBoard.scss
