+++ src/app/CompetitionGraph/CompetitionGraph.scss
