+++ src/app/CompetitionPools/CompetitionPools.scss
