+++ src/app/CompetitionSheet/CompetitionSheet.scss
