+++ src/app/CompetitionSheet/CompetitionSheet.tsx
@@ -3,8 +3,8 @@ import { useMemo, useState } from 'react';
 import { Layout, Result, Select, Space, Spin } from 'antd';
 import cx from 'classnames';
 
-import { Competition, Match, Sheet, Team } from '../../model/model';
-import { CSheetStat, CSStats } from '../../model/sheet';
+import { Competition, Match, Team } from '../../model/model';
+import { CSheetStat, CSStats, Sheet } from '../../model/sheet';
 import { acceptLicences, calcCSStats, filterPointSheets } from '../../model/sheet-helpers';
 import useSheets from '../../utils/useSheets';
