+++ src/app/CompetitionTeams/CompetitionTeams.scss
